requirements.txt
Django==5.1.8
djangorestframework==3.15.2
django-filter==24.3
djangorestframework-simplejwt==5.5.0
drf-spectacular==0.28.0
django-cors-headers==4.6.0
django-environ==0.11.2
Pillow==11.1.0

